Add a document explaining RPM tag values.
The LSB package standard doc stinks at actually explaining what goes in
an RPM, so here's an attempt to decipher it. This covers everything
specified in LSB-5.0.0, plus a few extras.

The file is markdown-style but the formatting is all screwed up, so it's
named .txt until that gets fixed.