Move the per-file tag handling out of mkrpm.c
Create a new function in tagdb.c that handles the per-file tags. This
function is called when adding a libarchive archive_entry to the cpio
file.

Add an extra check before calling add_file_tags to see if libarchive
unset the file size, since rpm expects the actual file size for the
header metadata.

Also, fix a bug in the per-file tag values. All of the integer tags need
to be in network byte order.