Fix issues with constructing the RPM headers.
construct_header was using a size of 32 as the size of the header and
index structures, which is double the actual size. Also, open_memstream
and fseek don't work like that: seeking backwards in the stream
effectively truncates it. The function can easily know the size of the
header ahead of time anyway, so just allocate a regular old chunk of
memory and write to it the normal way.